% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crashapi-package.R
\docType{package}
\name{crashapi-package}
\alias{crashapi}
\alias{crashapi-package}
\title{crashapi: CrashAPI}
\description{
Get Fatality Analysis Reporting System (FARS) data with the FARS API.
}
\keyword{internal}
