% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fars_terms}
\alias{fars_terms}
\title{NHSTA Terms and Definitions}
\format{
A data frame with 66 rows and 2 variables:
\describe{
\item{\code{term}}{character Term}
\item{\code{definition}}{character Term definition}
}
}
\source{
\href{https://www-fars.nhtsa.dot.gov/Help/Terms.aspx}{NHTSA FARS Terms}
}
\usage{
fars_terms
}
\description{
FARS-related terms defined by the National Highway Traffic
Safety Administration based on ANSI D16.1-1996: Manual on Classification of
Motor Vehicle Traffic Accidents.
}
\details{
Added: October 25 2021
Updated: October 25 2021
}
\keyword{datasets}
