% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_crashes}
\alias{format_crashes}
\title{Format crash data}
\usage{
format_crashes(x, details = TRUE)
}
\arguments{
\item{x}{Data frame with crash data.}

\item{details}{If \code{TRUE}, append date, time, datetime columns to formatted
crash data; defaults to TRUE}
}
\description{
Reorder columns to match the order documented in Fatality Analysis Reporting
System (FARS) Analytical User's Manual, 1975-2019 and append derived columns
for date, time, and datetime.
}
