% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_crss_zip.R
\name{get_crss_zip}
\alias{get_crss_zip}
\title{Download CRSS data files as zipped CSV or SAS files}
\usage{
get_crss_zip(
  year = 2021,
  format = "csv",
  path = NULL,
  aux = FALSE,
  read = TRUE,
  geometry = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{year}{Year of data from 2016 to 2021, Default: 2021}

\item{format}{Format of zipped data tables ('csv' or 'sas'). Default: 'csv'.
unzip and geo options are only supported if format is "csv".}

\item{path}{Path to download zip file. Set to \code{\link[=getwd]{getwd()}} if \code{NULL} (default).}

\item{aux}{If \code{TRUE}, download auxiliary
CRSS datasets .}

\item{read}{If \code{TRUE}, unzip the downloaded file and read CSV files into a
list of tables with each list item corresponding to one CSV file.}

\item{geometry}{If \code{TRUE}, convert the accident table to a sf object.}

\item{overwrite}{If \code{FALSE}, abort if file exists at the provided path. If
\code{TRUE}, overwrite file.}
}
\value{
Downloads zip file with CSV or SAS tables and returns the zip file
path invisibly or returns a list of data frames (if geo is \code{FALSE}), or
returns a list of data frames with the accident table converted to a sf
object.
}
\description{
This function is similar to \code{\link[=get_fars_zip]{get_fars_zip()}} to download files directly from
NHTSA FTP site. If read is \code{TRUE}, the function reads a list containing data
frames for each table available in the selected year. If geometry is \code{TRUE},
the accident table is converted to an sf object.
}
