% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars_crashes}
\alias{get_fars_crashes}
\alias{get_fars_crash_list}
\alias{get_fars_summary}
\title{Get Fatality Analysis Reporting System (FARS) data with the FARS API}
\usage{
get_fars_crashes(
  start_year = 2014,
  end_year = 2015,
  state = 1,
  county = 1,
  geometry = FALSE,
  crs = 4326
)

get_fars_crash_list(
  start_year = 2014,
  end_year = 2015,
  state = 1,
  vehicles = c(1, 2)
)

get_fars_summary(start_year = 2014, end_year = 2015, state = 1)
}
\arguments{
\item{start_year}{Required. Start year for crash reports (must be between
2010 and 2019), Default: 2014}

\item{end_year}{Required. End year for crash reports (must be between 2010
and 2019), Default: 2015}

\item{state}{Required. State name, abbreviation, or FIPS number.
\code{get_fars_crash_list} supports multiple states.}

\item{county}{County name or FIPS number. Required for \code{get_fars_crashes}.}

\item{geometry}{If TRUE, return sf object. Optional  for \code{get_fars_crashes}.}

\item{crs}{Coordinate reference system to return for \code{get_fars_crashes} if
geometry is TRUE}

\item{vehicles}{Vector with the minimum and maximum number of vehicles, e.g.
c(1, 2) for minimum of 1 vehicle and maximum of 2. Required for
\code{get_fars_crash_list}.}
}
\description{
Three APIs from FARS are currently supported:
\itemize{
\item \code{get_fars_crash_list} returns a list of fatal crashes that have occurred in
multiple States in one or more years. Limit the year range as data limit
has been set to 5000 records.
\item \code{get_fars_crashes} a list of fatal crashes by location that have occurred
throughout U.S. Limit the year range as data limit has been set to 5000
records.
\item \code{get_fars_summary} provides a count of injury severity that have occurred
throughout U.S. including count of fatalities and crashes.
}
}
