% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars_crash_persons}
\alias{get_fars_crash_persons}
\title{Get Crashes By Occupant}
\usage{
get_fars_crash_persons(
  year = NULL,
  start_year,
  end_year = NULL,
  state,
  age = NULL,
  sex = NULL,
  seat,
  injury,
  occupants = TRUE,
  nonoccupants = TRUE
)
}
\arguments{
\item{year}{numeric vector. Year or range with start and end year. 2010 to
2020 supported.}

\item{start_year}{Start year for crash reports.}

\item{end_year}{End year for crash reports.}

\item{state}{Required. State name, abbreviation, or FIPS number.}

\item{age}{numeric}

\item{sex}{Options "m", "f", "male", "female", "unknown", "not reported."}

\item{seat}{Seat position}

\item{injury}{Options "unknown", "not reported", "died prior", "injured",
"fatal", "suspected serious", "suspected minor", "possible", "no apparent"}

\item{occupants}{Include vehicle occupants in query; defaults to \code{TRUE}}

\item{nonoccupants}{Include non-occupants in query; defaults to \code{TRUE}}
}
\description{
This function returns a list of fatal crashes by occupant that have occurred
throughout United States.
}
