% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars_crash_vehicles}
\alias{get_fars_crash_vehicles}
\title{Get Get Crashes By Vehicle}
\usage{
get_fars_crash_vehicles(
  start_year = 2014,
  end_year = 2015,
  state = NULL,
  make = NULL,
  model = NULL,
  model_year = NULL,
  body_type = NULL
)
}
\arguments{
\item{start_year}{Required. Start year for crash reports (must be between
2010 and 2019), Default: 2014}

\item{end_year}{Required. End year for crash reports (must be between 2010
and 2019), Default: 2015}

\item{state}{Required. State name, abbreviation, or FIPS number.
\code{get_fars_crash_list} supports multiple states.}

\item{make}{Make name or ID, Required. The start_year is used to return a
list of support make options. Default: NULL}

\item{model}{Model name or ID, Optional. Default: NULL}

\item{model_year}{Model year, Optional. Default: NULL}

\item{body_type}{Body type, Optional. `model` must be provided to use
body_type parameter. Default: NULL}
}
\description{
This function returns a list of fatal crashes by vehicle type
  that have occurred throughout United States. The make, model, and body type
  must match the options returned by `fars_vars`. This function accepts named
  options that are converted to ID numbers for use in the API query.
}
