% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{list_rbind}
\alias{list_rbind}
\title{Combine list elements into a single data structure}
\usage{
list_rbind(x, ..., names_to = rlang::zap(), ptype = NULL)
}
\description{
\itemize{
\item \code{list_rbind()} combines elements into a data frame by row-binding them
together with \code{\link[vctrs:vec_bind]{vctrs::vec_rbind()}}.
}
}
\keyword{internal}
