% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mmucc_codes}
\alias{mmucc_codes}
\title{Model Minimum Uniform Crash Criteria (MMUCC) codes (simple)}
\format{
A data frame with 700 rows and 6 variables:
\describe{
\item{\code{code}}{Attribute code}
\item{\code{name}}{Attribute code name}
\item{\code{type}}{Attribute code type}
\item{\code{definition}}{Code definition}
\item{\code{restriction_id}}{Restriction id number}
\item{\code{restriction}}{Restriction value}
}
}
\usage{
mmucc_codes
}
\description{
A collection of the 73 unique codes identified as simple codes:
\url{https://release.niem.gov/niem/codes/mmucc/4.1/mmucc.xsd}

See the MMUCC Guideline Fifth Edition (2017) for more information:
\url{https://crashstats.nhtsa.dot.gov/Api/Public/Publication/812433}

About MMUC from NHTSA: \url{https://www.nhtsa.gov/mmucc-1}

To encourage greater uniformity, the National Highway Traffic Safety
Administration (NHTSA) and the Governors Highway Safety Association (GHSA)
cooperatively developed a voluntary data collection guideline in 1998. The
MMUCC guideline identifies a minimum set of motor vehicle crash data elements
and their attributes that States should consider collecting and including in
their State crash data system.

The MMUCC 5th Edition is the result of an 18-month collaboration between
NHTSA, the Federal Highway Administration (FHWA), the Federal Motor Carrier
Safety Administration (FMCSA), the National Transportation Safety Board
(NTSB), the GHSA, and subject matter experts from State DOTs, local law
enforcement, emergency medical services, safety organizations, industry
partners, and academia. The traffic records community and general public also
contributed through external forums (Federal Register) and at the 2016
Traffic Records Forum.
\itemize{
\item Added: March 26 2022
\item Updated: March 26 2022
}
}
\keyword{datasets}
