% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_crashapi}
\alias{read_crashapi}
\title{Read data from the CrashAPI}
\usage{
read_crashapi(
  url = "https://crashviewer.nhtsa.dot.gov",
  data = "crashes",
  type = NULL,
  format = "json",
  results = TRUE,
  ...
)
}
\arguments{
\item{url}{Base url for CrashAPI.}

\item{data}{Data (crashes, analytics, or fars), Default: 'crashes'}

\item{type}{Type of API to use, Default: NULL}

\item{format}{Format to return, Default: 'json'}

\item{results}{If FALSE, return formatted url, Default: TRUE}

\item{...}{Additional parameters used in template (varies by type).}
}
\value{
Data frame with requested data or a formatted url
}
\description{
An updated utility function using the httr2 package to read data from the CrashAPI.
}
