% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars}
\alias{get_fars}
\alias{get_fars_crashes}
\alias{get_fars_cases}
\alias{get_fars_crash_details}
\alias{get_fars_crash_list}
\alias{get_fars_summary}
\alias{get_fars_year}
\title{Get Fatality Analysis Reporting System (FARS) data with the FARS API}
\usage{
get_fars(
  year = 2022,
  state,
  county = NULL,
  api = c("crashes", "cases", "state list", "summary count", "year dataset", "zip"),
  type = NULL,
  details = FALSE,
  geometry = FALSE,
  crs = NULL,
  cases = NULL,
  vehicles = NULL,
  format = "json",
  pr = FALSE,
  path = NULL,
  download = FALSE
)

get_fars_crashes(
  year = 2022,
  start_year,
  end_year = NULL,
  state,
  county,
  details = FALSE,
  geometry = FALSE,
  crs = NULL
)

get_fars_cases(
  year = 2022,
  state,
  cases,
  details = FALSE,
  geometry = FALSE,
  crs = NULL
)

get_fars_crash_list(
  year = 2022,
  start_year = NULL,
  end_year = NULL,
  state,
  vehicles = c(1, 50)
)

get_fars_summary(year = 2022, start_year, end_year = NULL, state)

get_fars_year(
  year = 2022,
  type = "accident",
  state,
  format = "json",
  path = NULL,
  geometry = FALSE,
  crs = NULL,
  download = FALSE
)
}
\arguments{
\item{year}{numeric vector. Year or range with start and end year. If \code{api}
is "details", "year dataset", or "zip" (or using the
\code{get_fars_crash_details}, \code{get_fars_year}, or \code{get_fars_zip} functions), a
single year is required. All other \code{api} options support a range with the
minimum value is used as a start year and the maximum value used as a end
year. Most \code{api} options support the years from 2010 through the most
recent year of release. "year dataset" only supports 2010 to 2017 and "zip"
supports 1975 to 2022. \code{start_year} and \code{end_year} are ignored if \code{year} is
not \code{NULL}.}

\item{state}{Required. State name, abbreviation, or FIPS number.
\code{get_fars_crash_list} supports multiple states.}

\item{county}{County name or FIPS number. Required for \code{get_fars_crashes}.}

\item{api}{character. API function to use. Supported values include
"crashes", "cases", "state list", "summary count", "year dataset", and
"zip". Default: "crashes".}

\item{type}{Name of the dataset or data file to download when using the "year
dataset" api or \code{get_fars_year}. Supported values include "ACCIDENT",
"CEVENT", "DAMAGE", "DISTRACT", "DRIMPAIR", "FACTOR", "MANEUVER",
"NMCRASH", "NMIMPAIR", "NMPRIOR", "PARKWORK", "PBTYPE", "PERSON",
"SAFETYEQ", "VEHICLE", "VEVENT VINDECODE", "VINDERIVED", "VIOLATION",
"VISION", and "VSOE". Lowercase or mixed case values are permitted.}

\item{details}{Type of detailed crash data to return (either "events" or
"vehicles"). If \code{TRUE} for \code{get_fars} or \code{get_fars_crashes}, detailed case
data (excluding event and vehicle data) is attached to the returned crash
data. If \code{NULL} for \code{get_fars_cases}, events and vehicle data are excluded
from the returned case data. returned by \code{get_fars_cases}. Optional for
\code{get_fars_crash_details}. Default: \code{NULL} for \code{get_fars_cases}; \code{FALSE} for
\code{get_fars} and \code{get_fars_crashes}.}

\item{geometry}{If \code{TRUE}, return sf object. Optional for \code{get_fars_crashes}.}

\item{crs}{Coordinate reference system to return for \code{get_fars_crashes} if
\code{geometry} is \code{TRUE}.}

\item{cases}{One or more FARS case numbers. Required if \code{api} is "cases" or
using \code{get_fars_cases}. Multiple case numbers can be provided.}

\item{vehicles}{numeric vector with the minimum and maximum number of
vehicles, e.g. c(1, 2) for minimum of 1 vehicle and maximum of 2. Required
for \code{get_fars_crash_list}.}

\item{format}{Default "json". "csv" is supported when using the "year
dataset" api. "sas" is supporting for the "zip" api.}

\item{pr}{logical. If \code{TRUE}, download zip file with FARS data for Puerto
Rico. No Puerto Rico data available for years 1975-1977. Default: \code{FALSE}
for \code{get_fars_zip} only.}

\item{path}{File path used if download is \code{TRUE}.}

\item{download}{logical. If \code{TRUE} and the \code{api} is "year dataset" or "zip",
download the data to a file. Default \code{FALSE}.}

\item{start_year}{Start year for crash reports.}

\item{end_year}{End year for crash reports.}
}
\description{
This function provides a convenient interface for accessing FARS
data or data summaries using a range of criteria. The \code{api} parameter
allows you to call one of the following functions to access DOT NHTSA’s
Crash API:
\itemize{
\item \code{get_fars_crash_list} returns a list of fatal crashes that have occurred
in multiple states in one or more years.
\item \code{get_fars_crash_details} returns a details of a fatal crash that has
occurred in a state for a single year.
\item \code{get_fars_crashes} a list of fatal crashes by location that have occurred
throughout U.S.
\item \code{get_fars_summary} provides a count of injury severity that have occurred
throughout U.S. including count of fatalities and crashes.
\item \code{get_fars_year} provides one of 20 FARS data tables for a single year.
Supports downloading to a CSV or JSON file.
}

Both \code{get_fars_crash_list} and \code{get_fars_crashes} limit the returned data
to 5000 records so consider limiting the range of years requested if data
exceeds that threshold.

This package also enables access to the FARS data available through the
NHTSA data downloads server in a zip format. Set \code{api} to "zip" or use the
\code{get_fars_zip} function to download this data.
}
\examples{

head(get_fars_crashes(state = "MD", county = "Baltimore city"), 5)

get_fars_cases(state = "MD", cases = "240274")

get_fars_crash_list(state = "MD", vehicles = 5)

get_fars_summary(state = "MD")

head(get_fars_year(state = "MD", type = "PERSON"), 5)

}
