% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars_crash_vehicles}
\alias{get_fars_crash_vehicles}
\title{Get Crashes By Vehicle}
\usage{
get_fars_crash_vehicles(
  year = NULL,
  start_year,
  end_year,
  state,
  make = NULL,
  model = NULL,
  model_year = 2010,
  body_type = NULL
)
}
\arguments{
\item{year}{numeric vector. Year or range with start and end year. 2010 to 2019 supported.}

\item{start_year}{Start year for crash reports.}

\item{end_year}{End year for crash reports.}

\item{state}{Required. State name, abbreviation, or FIPS number.}

\item{make}{Make name or ID, Required. The start_year is used to return a
list of support make options. Default: NULL}

\item{model}{Model name or ID, Optional. Default: NULL}

\item{model_year}{Model year, Optional. Default: NULL}

\item{body_type}{Body type, Optional. `model` must be provided to use
body_type parameter. Default: NULL}
}
\description{
This function returns a list of fatal crashes by vehicle type
  that have occurred throughout United States. The make, model, and body type
  must match the options returned by `fars_vars`. This function accepts named
  options that are converted to ID numbers for use in the API query.
}
