% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars.R
\name{get_fars_zip}
\alias{get_fars_zip}
\title{Download FARS data files as zipped CSV or SAS files}
\usage{
get_fars_zip(year = 2019, format = "csv", path = NULL, pr = FALSE)
}
\arguments{
\item{year}{Year of data from 1975 to 2019, Default: 2019}

\item{format}{Format of zipped data tables ('csv' or 'sas'). Default: 'csv'.}

\item{path}{Path to download zip file with FARS tables.}

\item{pr}{If TRUE, download FARS data for Puerto Rico. No Puerto Rico data available for years 1975-1977. Default: FALSE}
}
\value{
Downloads zip file with CSV or SAS tables.
}
\description{
This function provides an alternative to get_fars_year that downloads files directly from NHTSA FTP site.
}
