% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_fars_zip.R
\name{get_fars_zip}
\alias{get_fars_zip}
\title{Download FARS data files as zipped CSV or SAS files}
\usage{
get_fars_zip(
  year = 2023,
  format = "csv",
  path = NULL,
  pr = FALSE,
  aux = FALSE,
  read = TRUE,
  geometry = FALSE,
  overwrite = FALSE
)
}
\arguments{
\item{year}{Year of data from 1975 to 2024, Default: 2024}

\item{format}{Format of zipped data tables ('csv' or 'sas'). Default: 'csv'.
unzip and geo options are only supported if format is "csv".}

\item{path}{Path to download zip file. Set to \code{\link[=getwd]{getwd()}} if \code{NULL} (default).}

\item{pr}{If \code{TRUE}, download FARS data for Puerto Rico. No Puerto Rico data
available for years 1975-1977. Default: \code{FALSE}}

\item{aux}{If \code{TRUE} and year is after 1982, download auxiliary
FARS datasets that "contain data derived from the regular FARS/GES
variables using NCSA analytical data classifications." In 2010, the NHTSA
explained: "These classifications are widely used in NCSA publications and
research. Many definitions such as "speeding-related" or "distracted
driving" comprise a combination of variables whose names or attributes have
changed over time. The derived variables in the auxiliary files incorporate
these nuances, thus simplifying the use of standard classifications in any
traffic safety research." Learn more from the FARS and GES Auxiliary
Datasets Q & A:
\url{https://crashstats.nhtsa.dot.gov/Api/Public/ViewPublication/811364}}

\item{read}{If \code{TRUE}, unzip the downloaded file and read CSV files into a
list of tables with each list item corresponding to one CSV file.}

\item{geometry}{If \code{TRUE}, convert the accident table to a sf object.}

\item{overwrite}{If \code{FALSE}, abort if file exists at the provided path. If
\code{TRUE}, overwrite file.}
}
\value{
Downloads zip file with CSV or SAS tables and returns the zip file
path invisibly or returns a list of data frames (if geometry is \code{FALSE}),
or returns a list of data frames with the accident table converted to a sf
object.
}
\description{
This function provides an alternative to \code{\link[=get_fars_year]{get_fars_year()}} that downloads
files directly from NHTSA FTP site. If read is \code{TRUE}, the function reads a
list containing data frames for each table available in the selected year. If
geometry is \code{TRUE}, the accident table is converted to an sf object.
}
