% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_crashapi}
\alias{read_crashapi}
\title{Read data from the CrashAPI using a url template}
\usage{
read_crashapi(
  url = "https://crashviewer.nhtsa.dot.gov",
  data = "crashes",
  type = NULL,
  format = "json",
  results = TRUE,
  ...,
  call = caller_env()
)
}
\arguments{
\item{url}{Base url for CrashAPI.}

\item{data}{Data (crashes, analytics, or fars), Default: 'crashes'}

\item{type}{Type of API to use, Default: \code{NULL}}

\item{format}{Format to return, Default: 'json'}

\item{results}{If \code{FALSE}, return formatted url, Default: \code{TRUE}}

\item{...}{Additional parameters used in template (varies by type).}
}
\value{
Data frame with requested data or a formatted url (if \code{results = FALSE})
}
\description{
An updated utility function using the httr2 package to read data from the
CrashAPI using the API URL templates listed on the NHSTA website:
\url{https://crashviewer.nhtsa.dot.gov/CrashAPI}
}
