% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_crashapi}
\alias{read_crashapi}
\title{Read data from the CrashAPI using a url template}
\usage{
read_crashapi(
  url = "https://crashviewer.nhtsa.dot.gov/crashviewer/CrashAPI",
  data = "crashes",
  type = NULL,
  format = "json",
  results = TRUE,
  ...,
  cookie_file = "crashapi.cookies",
  call = caller_env()
)
}
\arguments{
\item{url}{Base url for CrashAPI.}

\item{data}{Data (crashes, analytics, or fars), Default: 'crashes'}

\item{type}{Type of API to use, Default: \code{NULL}}

\item{format}{Format to return, Default: 'json'}

\item{results}{If \code{FALSE}, return formatted url, Default: \code{TRUE}}

\item{...}{Additional parameters used in template (varies by type).}

\item{call}{The execution environment of a currently
running function, e.g. \code{caller_env()}. The function will be
mentioned in error messages as the source of the error. See the
\code{call} argument of \code{\link[rlang:abort]{abort()}} for more information.}
}
\value{
Data frame with requested data or a formatted url (if \code{results = FALSE})
}
\description{
An updated utility function using the httr2 package to read data from the
CrashAPI using the API URL templates listed on the NHSTA website:
\url{https://crashviewer.nhtsa.dot.gov/crashviewer/CrashAPI}
}
